我也忘记改了啥……自动生成了一堆比赛表的pojo，加入mp5s6和mp4s3的进群公告

+++ src/main/java/top/mothership/cabbage/controller/vo/ChartsVo.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 @Data
 public class ChartsVo {
-    private List<Integer> xAxis;
+    private List<Long> xAxis;
     private List<Integer> yAxis;
 }
